cleanups

+++ pimux2000/pimux2000App.swift
@@ -9,11 +9,22 @@ struct pimux2000App: App {
 
 	init() {
 		let processInfo = ProcessInfo.processInfo
-		let databaseURL = URL.documentsDirectory.appending(path: "db.sqlite")
+
+		if Self.isRunningForPreviews {
+			appDatabase = AppDatabase.preview()
+			return
+		}
+
+		let databaseURL = Self.databaseURL
 
 		if processInfo.arguments.contains("--uitesting-reset-db") {
 			try? FileManager.default.removeItem(at: databaseURL)
 		}
+
+		try! FileManager.default.createDirectory(
+			at: databaseURL.deletingLastPathComponent(),
+			withIntermediateDirectories: true
+		)
 
 		appDatabase = try! AppDatabase(
 			dbQueue: DatabaseQueue(path: databaseURL.path())
@@ -30,5 +41,13 @@ struct pimux2000App: App {
 				.environment(\.appDatabase, appDatabase)
 				.databaseContext(.readWrite { appDatabase.dbQueue })
 		}
+	}
+
+	private static var isRunningForPreviews: Bool {
+		ProcessInfo.processInfo.environment["XCODE_RUNNING_FOR_PREVIEWS"] == "1"
+	}
+
+	private static var databaseURL: URL {
+		URL.documentsDirectory.appending(path: "db.sqlite")
 	}
 }
